backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==12.0
Appium-Python-Client==4.0.0
google-generativeai==0.8.0
Pillow==10.4.0
aiofiles==24.1.0
python-multipart==0.0.9
pydantic==2.9.0
imagehash==4.3.1
